+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsetypescript.py
@@ -186,9 +186,9 @@ class EvaluatorResponseTypescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-01T08:24:55.375Z"
+    created: Optional[str] = "2026-07-01T08:41:13.115Z"
 
-    updated: Optional[str] = "2026-07-01T08:24:55.376Z"
+    updated: Optional[str] = "2026-07-01T08:41:13.115Z"
 
     updated_by_id: OptionalNullable[str] = UNSET