+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWECGYV7WY3E2SX3CA4MX19Y"
+        "01KWEDEPTC4WNXVN1A2S59948D"
     )
     r"""The unique identifier of the data source"""