+++ packages/orq-rc/RELEASES.md
@@ -9129,3 +9129,13 @@ Based on:
 - [python v4.11.0-rc.63] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.63] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.63 - packages/orq-rc
+
+## 2026-07-01 08:43:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.787.0 (2.914.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.64] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.64] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.64 - packages/orq-rc